// dashboard/src/utils/hasura-api/generated/schemas/getScheduledEventsArgs.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { GetScheduledEventsArgsType } from './getScheduledEventsArgsType';
import type { ScheduledEventStatus } from './scheduledEventStatus';

export interface GetScheduledEventsArgs {
  /** Type of the Scheduled Trigger. */
  type: GetScheduledEventsArgsType;
  /** Name of the cron trigger. Required only when type is cron. */
  trigger_name?: string;
  /** Maximum number of events to be returned in one API call. */
  limit?: number;
  /** Starting point from where the rows need to be returned. */
  offset?: number;
  /** Flag to indicate whether number of rows returned should be included in the response. */
  get_rows_count?: boolean;
  /** Filter scheduled events by the status */
  status?: ScheduledEventStatus[];
}