// dashboard/src/utils/hasura-api/generated/schemas/getScheduledEventsResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ScheduledEventLogEntry } from './scheduledEventLogEntry';

export interface GetScheduledEventsResponse {
  events: ScheduledEventLogEntry[];
}